requirements.txt
networkx==3.2.1
matplotlib==3.8.2
pyyaml==6.0.1
rich==13.7.0
flask==3.0.0
requests==2.31.0
pandas==2.1.4
